pyproject.toml
[tool.poetry]
name        = "VISY-REVE-PY"
version     = "0.1.0"
description = ""
readme      = "README.md"
authors     = ["Marius Neuhalfen <neuhalfen.marius@gmail.com>"]

packages = [
    { include = "training" }
]

[tool.poetry.dependencies]
python                        = ">=3.11"
asttokens                     = "3.0.0"
attrs                         = "25.3.0"
blinker                       = "1.9.0"
certifi                       = "2025.4.26"
charset-normalizer            = "3.4.2"
click                         = "8.2.1"
colorama                      = "0.4.6"
comm                          = "0.2.2"
ConfigArgParse                = "1.7.1"
contourpy                     = "1.3.2"
cycler                        = "0.12.1"
dash                          = "3.0.4"
decorator                     = "5.2.1"
executing                     = "2.2.0"
fastjsonschema               = "2.21.1"
Flask                         = "3.0.3"
fonttools                     = "4.58.0"
freetype-py                   = "2.5.1"
idna                          = "3.10"
imageio                       = "2.37.0"
importlib-metadata            = "8.7.0"
ipython                       = "9.2.0"
ipython-pygments-lexers       = "1.1.1"
ipywidgets                    = "8.1.7"
itsdangerous                  = "2.2.0"
jedi                          = "0.19.2"
Jinja2                        = "3.1.6"
jsonschema                    = "4.23.0"
jsonschema-specifications     = "2025.4.1"
jupyter-core                  = "5.7.2"
jupyterlab-widgets            = "3.0.15"
kiwisolver                    = "1.4.8"
MarkupSafe                    = "3.0.2"
matplotlib                    = "3.10.3"
matplotlib-inline             = "0.1.7"
narwhals                      = "1.40.0"
nbformat                      = "5.10.4"
nest-asyncio                  = "1.6.0"
networkx                      = "3.4.2"
numpy                         = "2.2.6"
open3d                        = "0.19.0"
opencv-python                 = ">=4.10.0.84"
packaging                     = "25.0"
pandas                        = "2.2.3"
parso                         = "0.8.4"
pillow                        = "11.2.1"
platformdirs                  = "4.3.8"
plotly                        = "6.1.1"
prompt-toolkit                = "3.0.51"
pure-eval                     = "0.2.3"
pyglet                        = "2.1.6"
Pygments                      = "2.19.1"
PyOpenGL                      = "3.1.0"
pyparsing                     = "3.2.3"
pypng                         = "0.20220715.0"
pyrender                      = "0.1.45"
python-dateutil               = "2.9.0.post0"
pytz                          = "2025.2"
pywin32                       = "310"
referencing                   = "0.36.2"
requests                      = "2.32.3"
retrying                      = "1.3.4"
rpds-py                       = "0.25.1"
scipy                         = "1.15.3"
six                           = "1.17.0"
stack-data                    = "0.6.3"
traitlets                     = "5.14.3"
trimesh                       = "4.6.10"
typing-extensions             = "4.13.2"
tzdata                        = "2025.2"
urllib3                       = "2.4.0"
wcwidth                       = "0.2.13"
Werkzeug                      = "3.0.6"
widgetsnbextension            = "4.0.14"
zipp                          = "3.21.0"

# Git dependency for cv-utils
cv-utils-mariusn = { git = "https://github.com/marius-ne/CV_utils.git" }
scikit-image = "^0.25.2"
largestinteriorrectangle = "^0.2.1"
seaborn = "^0.13.2"
ipykernel = "^6.29.5"
scikit-learn = "^1.6.1"
torch = "^2.7.0"
torchvision = "^0.22.0"
yacs = "^0.1.8"
json-tricks = "^3.17.3"
pycocotools = "^2.0.9"

[build-system]
requires    = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"
